pyproject.toml
[project]
name = "project-readiness-checker"
version = "0.1.0"
description = "A tool to verify project readiness based on configurable rules."
authors = [
    { name = "Robert Esteban" }
]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0"
]
requires-python = ">=3.8"

[project.scripts]
prc = "src.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
